requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
chardet==3.0.4
click==7.1.1
discord.py==1.3.2
entrypoints==0.3
flake8==3.7.9
idna==2.9
isort==4.3.21
mccabe==0.6.1
multidict==4.7.5
pathspec==0.7.0
psutil==5.7.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.12.0
regex==2020.2.20
toml==0.10.0
typed-ast==1.4.1
websockets==8.1
yarl==1.4.2
